+++ .husky/prepare-commit-msg
@@ -1,0 +1,4 @@
+#!/bin/sh
+# gitmoji as a commit hook
+exec < /dev/tty
+gitmoji --hook $1 $2
