+++ src/core/cashback/use-cases/validate-cashback.ts
