+++ src/core/product/adapter/product.ts
