+++ src/core/product/types/product.ts
